Fix "small monospace fonts" workaround to work properly
It's true that to get monospaced fonts to size right, you
need to munge the font fallback list.

http://neugierig.org/software/chromium/notes/2009/09/monospace-fonts-workaround.html

But when a font named "monospace" isn't available, Chrome will fall back
on the next option in the list (serif, in this code), which results in a
serif font for <pre> blocks.

Instead, put a nonexistent font *first* so that it always falls
back to monospace.